Switch to `getopt` for option parsing
Since it is easy to do so now, add a `-h` option that prints a usage message
and switch the chroot functionality to be specified via the `-c` option.

More options are to come soon, so building the machinery of a more
robust `usage` function and `parse_arguments` function is useful.

Signed-off-by: Charles Mirabile <cmirabil@redhat.com>