switch order of chroot and chdir to simplify avoiding race conditions
no matter which order it is performed there will be a breif window of
incongruiety between /proc/self/root and /proc/self/cwd. Perhaps cwd
being outside of root is less intuitive and worse in a threaded program
but for this single threaded application doing things in this order is
cleaner.

The prior implementation leaked the file descriptor opened for the
path, and worse, because it forgot to set cloexec, it was inherited
by the handler program.

This option does not require the fuss with /proc and still avoids a
potential race condition because chroot performs path resolution once
and after it returns the abs path `"/"` will refer to whatever path it
settled on.